Fix kube agent shutdown during upgrades (#21616)
During binary upgrades, Teleport forks the agents and calls the shutdown
method for each one of them.

Kubernetes Agent was calling the `http.Server` shutdown method and which
wasn't stopping the heartbeats neither handling the listeners.
This PR fixes the shutdown behaviour for Kube agent and avoids deleting
the heartbeated resources when the agent is reloading.

+++ lib/kube/proxy/watcher.go
@@ -208,13 +208,24 @@ func (s *TLSServer) updateKubeCluster(ctx context.Context, cluster types.KubeClu
 	return nil
 }
 
+// unregisterKubeCluster unregisters the proxied Kube Cluster from the agent.
+// This function is called when the dynamic cluster is deleted/no longer match
+// the agent's resource matcher or when the agent is shutting down.
 func (s *TLSServer) unregisterKubeCluster(ctx context.Context, name string) error {
 	var errs []error
 
 	errs = append(errs, s.stopHeartbeat(name))
 	s.fwd.removeKubeDetails(name)
-	errs = append(errs, s.deleteKubernetesServer(ctx, name))
 
+	// A child process can be forked to upgrade the Teleport binary. The child
+	// will take over the heartbeats so do NOT delete them in that case.
+	// When unregistering a dynamic cluster, the context is empty and the
+	// decision will be to delete the kubernetes server.
+	if services.ShouldDeleteServerHeartbeatsOnShutdown(ctx) {
+		errs = append(errs, s.deleteKubernetesServer(ctx, name))
+	}
+
+	// close active sessions before returning.
 	s.fwd.mu.Lock()
 	sessions := maps.Values(s.fwd.sessions)
 	s.fwd.mu.Unlock()
